+++ dbatools.schema/stored_procs.sql
@@ -99,3 +99,64 @@ IF q_proc IS NULL OR q_proc = '' THEN
 END$$
 DELIMITER ;
 -- END
+
+-- PROCEDURE TO REPORT ON ENGINE USAGE --
+DROP PROCEDURE IF EXISTS `dbatools`.`ENGINE_REPORT`;
+DELIMITER $$
+CREATE DEFINER=`root`@`localhost` PROCEDURE `ENGINE_REPORT`()
+  COMMENT 'Generates a report for table engine usage (MyISAM, INNODB, MEMORY).'
+proc_label:BEGIN
+
+SELECT
+-- TOTALS
+ROUND(((data_size + index_size) / gb),4) AS total_size_gb,
+ROUND((index_size / gb),4) AS index_size_gb,
+ROUND((data_size / gb),4) AS data_size_gb,
+ROUND((index_size / (data_size + index_size)),2) * 100 AS perc_index,
+ROUND((data_size / (data_size + index_size)),2) * 100 AS perc_data,
+--
+-- InnoDB ENGINE REPORTING
+ROUND((SUM(innodb_index_size + innodb_data_size) / gb),4) AS innodb_total_size_gb,
+ROUND((innodb_data_size / gb),4) AS innodb_data_size_gb,
+ROUND((innodb_index_size / gb),4) AS innodb_index_size_gb,
+ROUND(innodb_index_size / (innodb_data_size + innodb_index_size),2) * 100 AS innodb_perc_index,
+ROUND(innodb_data_size / (innodb_data_size + innodb_index_size),2) * 100 AS innodb_perc_data,
+ROUND(innodb_index_size / index_size,2) * 100 AS innodb_perc_total_index,
+ROUND(innodb_data_size / data_size,2) * 100 AS innodb_perc_total_data,
+--
+-- MyISAM ENGINE REPORTING
+ROUND((SUM(myisam_index_size + myisam_data_size) / gb),4) AS myisam_total_size_gb,
+ROUND((myisam_data_size / gb),4) AS myisam_data_size_gb,
+ROUND((myisam_index_size / gb),4) AS myisam_index_size_gb,
+ROUND(myisam_index_size / (myisam_data_size + myisam_index_size),2) * 100 AS myisam_perc_index,
+ROUND(myisam_data_size / (myisam_data_size + myisam_index_size),2) * 100 AS myisam_perc_data,
+ROUND(myisam_index_size / index_size,2) * 100 AS myisam_perc_total_index,
+ROUND(myisam_data_size / data_size,2) * 100 AS myisam_perc_total_data,
+--
+-- MEMORY ENGINE REPORTING
+ROUND((SUM(memory_index_size + memory_data_size) / gb),4) AS memory_total_size_gb,
+ROUND((memory_data_size / gb),4) AS memory_data_size_gb,
+ROUND((memory_index_size / gb),4) AS memory_index_size_gb,
+ROUND(memory_index_size / (memory_data_size + memory_index_size),2) * 100 AS memory_perc_index,
+ROUND(memory_data_size / (memory_data_size + memory_index_size),2) * 100 AS memory_perc_data,
+ROUND(memory_index_size / index_size,2) * 100 AS memory_perc_total_index,
+ROUND(memory_data_size / data_size,2) * 100 AS memory_perc_total_data
+--
+-- Generate summaries from which to analyze
+FROM (
+  SELECT
+  SUM(data_length) data_size,
+  SUM(index_length) index_size,
+  SUM(if(engine = 'innodb', data_length, 0)) AS innodb_data_size,
+  SUM(if(engine = 'innodb', index_length, 0)) AS innodb_index_size,
+  SUM(if(engine = 'myisam', data_length, 0)) AS myisam_data_size,
+  SUM(if(engine = 'myisam', index_length, 0)) AS myisam_index_size,
+  SUM(if(engine = 'memory', data_length, 0)) AS memory_data_size,
+  SUM(if(engine = 'memory', index_length, 0)) AS memory_index_size,
+  POW(1024, 3) gb
+  FROM information_schema.tables
+  WHERE table_type = 'BASE TABLE') a;
+
+END$$
+DELIMITER ;
+-- END