+++ scripts/task.extract-single-table-from-backup.sh
@@ -1,0 +1,1 @@
+sed -n -e '/DROP TABLE.*'mytable'/,/UNLOCK TABLES/p’ backup.sql > mytable.sql