+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+- added engine_repor() proc to dbatools schema
+
+## 0.0.2
+- adding schema backup script and memory equations txt file
+
 ## 0.0.1
 - added PROC_LIST_ALL, PROC_LIST_SCHEMA, PROC_SEARCH, TRIGGER_LIST_ALL, TRIGGER_LIST_SCHEMA, TRIGGER_SEARCH to dbatools schema